Remove WebAPI typeclass, merge some modules
Currently for making a prototype typeclasses are not needed,
here they just make program structure more complicated and
only one instance can exist (for main implementation).
Better return them later for testing (or use Handle pattern).
